Add the inline command to prepare the apolloscape dataset
