+++ monstereo/prep/prep_apolloscape.py
@@ -1,8 +1,8 @@
-# pylint: disable=too-many-statements, import-error
+#!/usr/bin/env python
+# coding: utf-8
 
+# In[1]:
 
-"""Extract joints annotations and match with nuScenes ground truths
-"""
 
 import os
 import sys
@@ -13,208 +13,45 @@ import json
 import logging
 from collections import defaultdict
 import datetime
-
 import numpy as np
-from pyquaternion import Quaternion
+import torch
+from ..utils import append_cluster, correct_angle, normalize_hwl, pixel_to_camera, to_spherical
 
-from ..utils import get_iou_matches, append_cluster, select_categories, project_3d, correct_angle, normalize_hwl, \
-    to_spherical
-from ..network.process import preprocess_pifpaf, preprocess_monoloco
+from ..network.process import preprocess_monoloco
 
-from ..utils import K, car_name2id, car_id2name, intrinsic_vec_to_mat, euler_angles_to_rotation_matrix, \
-                                rotation_matrix_to_euler_angles, convert_pose_mat_to_6dof, project, pose_extraction
-
-class PreprocessApolloscape:
-
-    #reprocessNuscenes(args.dir_ann, args.dir_nuscenes, args.dataset, args.iou_min)
-    #prep.run()
-    """Preprocess apolloscape dataset"""
-    #AV_W = 0.68
-    #AV_L = 0.75
-    #AV_H = 1.72
-    #WLH_STD = 0.1
-    #social = False
-
-    #CAMERAS = ('CAM_FRONT', 'CAM_FRONT_LEFT', 'CAM_FRONT_RIGHT', 'CAM_BACK', 'CAM_BACK_LEFT', 'CAM_BACK_RIGHT')
-    dic_jo = {'train': dict(X=[], Y=[], names=[], kps=[], boxes_3d=[], K=[],
-                            clst=defaultdict(lambda: defaultdict(list))),
-              'val': dict(X=[], Y=[], names=[], kps=[], boxes_3d=[], K=[],
-                          clst=defaultdict(lambda: defaultdict(list))),
-              }
-    dic_names = defaultdict(lambda: defaultdict(list))
-
-    def __init__(self, dir_ann, dir_apollo, dataset, iou_min):
-
-        logging.basicConfig(level=logging.INFO)
-        self.logger = logging.getLogger(__name__)
-
-        self.iou_min = iou_min
-        self.dir_ann = dir_ann
-        dir_out = os.path.join('data', 'arrays')
-        assert os.path.exists(dir_apollo), "apollo directory does not exists"
-        assert os.path.exists(self.dir_ann), "The annotations directory does not exists"
-        assert os.path.exists(dir_out), "Joints directory does not exists"
-
-        now = datetime.datetime.now()
-        now_time = now.strftime("%Y%m%d-%H%M")[2:]
-        self.path_joints = os.path.join(dir_out, 'joints-' + dataset + '-' + now_time + '.json')
-        self.path_names = os.path.join(dir_out, 'names-' + dataset + '-' + now_time + '.json')
-
-        self.apollo, self.scenes, self.split_train, self.split_val = factory(dataset, dir_apollo)
-
-    def run(self):
-        """
-        Prepare arrays for training
-        """
-        cnt_scenes = cnt_samples = cnt_sd = cnt_ann = 0
-        start = time.time()
-        
-        for ii, scene in enumerate(self.scenes):
-            end_scene = time.time()
-            current_token = scene['first_sample_token']
-            cnt_scenes += 1
-            time_left = str((end_scene - start_scene) / 60 * (len(self.scenes) - ii))[:4] if ii != 0 else "NaN"
-
-            sys.stdout.write('\r' + 'Elaborating scene {}, remaining time {} minutes'
-                             .format(cnt_scenes, time_left) + '\t\n')
-            start_scene = time.time()
-            if scene['name'] in self.split_train:
-                phase = 'train'
-            elif scene['name'] in self.split_val:
-                phase = 'val'
-            else:
-                print("phase name not in training or validation split")
-                continue
-
-            while not current_token == "":
-                sample_dic = self.nusc.get('sample', current_token)
-                cnt_samples += 1
-
-                # Extract all the sample_data tokens for each sample
-                for cam in self.CAMERAS:
-                    sd_token = sample_dic['data'][cam]
-                    cnt_sd += 1
-
-                    # Extract all the annotations of the person
-                    path_im, boxes_obj, kk = self.nusc.get_sample_data(sd_token, box_vis_level=1)  # At least one corner
-                    boxes_gt, boxes_3d, ys = extract_ground_truth(boxes_obj, kk)
-                    kk = kk.tolist()
-                    name = os.path.basename(path_im)
-                    basename, _ = os.path.splitext(name)
-
-                    self.dic_names[basename + '.jpg']['boxes'] = copy.deepcopy(boxes_gt)
-                    self.dic_names[basename + '.jpg']['ys'] = copy.deepcopy(ys)
-                    self.dic_names[basename + '.jpg']['K'] = copy.deepcopy(kk)
-
-                    # Run IoU with pifpaf detections and save
-                    path_pif = os.path.join(self.dir_ann, name + '.pifpaf.json')
-                    exists = os.path.isfile(path_pif)
-
-                    if exists:
-                        with open(path_pif, 'r') as file:
-                            annotations = json.load(file)
-                            boxes, keypoints = preprocess_pifpaf(annotations, im_size=(1600, 900))
-                    else:
-                        continue
-
-                    if keypoints:
-                        matches = get_iou_matches(boxes, boxes_gt, self.iou_min)
-                        for (idx, idx_gt) in matches:
-                            keypoint = keypoints[idx:idx + 1]
-                            inp = preprocess_monoloco(keypoint, kk).view(-1).tolist()
-                            lab = ys[idx_gt]
-                            lab = normalize_hwl(lab)
-                            self.dic_jo[phase]['kps'].append(keypoint)
-                            self.dic_jo[phase]['X'].append(inp)
-                            self.dic_jo[phase]['Y'].append(lab)
-                            self.dic_jo[phase]['names'].append(name)  # One image name for each annotation
-                            self.dic_jo[phase]['boxes_3d'].append(boxes_3d[idx_gt])
-                            append_cluster(self.dic_jo, phase, inp, lab, keypoint)
-                            cnt_ann += 1
-                            sys.stdout.write('\r' + 'Saved annotations {}'.format(cnt_ann) + '\t')
-
-                current_token = sample_dic['next']
-
-        with open(os.path.join(self.path_joints), 'w') as f:
-            json.dump(self.dic_jo, f)
-        with open(os.path.join(self.path_names), 'w') as f:
-            json.dump(self.dic_names, f)
-        end = time.time()
-
-        extract_box_average(self.dic_jo['train']['boxes_3d'])
-        print("\nSaved {} annotations for {} samples in {} scenes. Total time: {:.1f} minutes"
-              .format(cnt_ann, cnt_samples, cnt_scenes, (end-start)/60))
-        print("\nOutput files:\n{}\n{}\n".format(self.path_names, self.path_joints))
+from ..utils import K, KPS_MAPPING,car_name2id, car_id2name, intrinsic_vec_to_mat, intrinsic_mat_to_vec, euler_angles_to_rotation_matrix,                                 rotation_matrix_to_euler_angles, convert_pose_mat_to_6dof, project, pose_extraction
 
 
-def extract_ground_truth(boxes_obj, kk, spherical=True):
 
-    boxes_gt = []
-    boxes_3d = []
-    ys = []
+#dataset = "train"
 
-    for box_obj in boxes_obj:
+#annotations = False
 
-        # Select category
-        if box_obj.name[:6] != 'animal':
-            general_name = box_obj.name.split('.')[0] + '.' + box_obj.name.split('.')[1]
-        else:
-            general_name = 'animal'
-        if general_name in select_categories('all'):
-
-            # Obtain 2D & 3D box
-            boxes_gt.append(project_3d(box_obj, kk))
-            boxes_3d.append(box_obj.center.tolist() + box_obj.wlh.tolist())
-
-            # Angle
-            yaw = quaternion_yaw(box_obj.orientation)
-            assert - math.pi <= yaw <= math.pi
-            sin, cos, _ = correct_angle(yaw, box_obj.center)
-            hwl = [float(box_obj.wlh[i]) for i in (2, 0, 1)]
-
-            # Spherical coordinates
-            xyz = list(box_obj.center)
-            dd = np.linalg.norm(box_obj.center)
-            if spherical:
-                rtp = to_spherical(xyz)
-                loc = rtp[1:3] + xyz[2:3] + rtp[0:1]  # [theta, psi, z, r]
-            else:
-                loc = xyz + [dd]
-
-            output = loc + hwl + [sin, cos, yaw]
-            ys.append(output)
-
-    return boxes_gt, boxes_3d, ys
+#kps_3d = False
+#buffer = 20
+#radius = 200
+#dir_ann ="/data/maxime-data/apollo-pifpaf/annotations/"
+#dir_apollo ="/data/maxime-data/apolloscape"
+#dir_out = "/home/maximebonnesoeur/monstereo/data/arrays/"
 
 
-def factory(dataset, dir_nuscenes):
-    """Define dataset type and split training and validation"""
-
-    assert dataset in ['train']
-
-    if dataset == 'train':
-        with open("splits/nuscenes_teaser_scenes.txt", "r") as file:
-            teaser_scenes = file.read().splitlines()
-        scenes = [scene for scene in scenes if scene['token'] in teaser_scenes]
-        with open("splits/split_nuscenes_teaser.json", "r") as file:
-            dic_split = json.load(file)
-        split_train = [scene['name'] for scene in scenes if scene['token'] in dic_split['train']]
-        split_val = [scene['name'] for scene in scenes if scene['token'] in dic_split['val']]
-    else:
-        split_scenes = splits.create_splits_scenes()
-        split_train, split_val = split_scenes['train'], split_scenes['val']
-
-    return nusc, scenes, split_train, split_val
+# In[6]:
 
 
-def quaternion_yaw(q: Quaternion, in_image_frame: bool = True) -> float:
-    if in_image_frame:
-        v = np.dot(q.rotation_matrix, np.array([1, 0, 0]))
-        yaw = -np.arctan2(v[2], v[0])
-    else:
-        v = np.dot(q.rotation_matrix, np.array([1, 0, 0]))
-        yaw = np.arctan2(v[1], v[0])
-    return float(yaw)
+def bbox_3d_extract(vertices):
+    min_x = np.min(vertices[:,0])
+    max_x= np.max(vertices[:,0])
+    min_y = np.min(vertices[:,1])
+    max_y= np.max(vertices[:,1])
+    min_z = np.min(vertices[:,2])
+    max_z= np.max(vertices[:,2])
+    
+    bbox_3d=[]
+    for x in [min_x, max_x]:
+        for y in [min_y, max_y]:
+            for z in [min_z, max_z]:
+                bbox_3d.append([x,y,z])
+    return(np.array(bbox_3d))
 
 
 def extract_box_average(boxes_3d):
@@ -225,55 +62,608 @@ def extract_box_average(boxes_3d):
     print(stds)
 
 
-def extract_social(inputs, ys, keypoints, idx, matches):
-    """Output a (padded) version with all the 5 neighbours
-    - Take the ground feet and the output z
-    - make relative to the person (as social LSTM)"""
-    all_inputs = []
-
-    # Find the lowest relative ground foot
-    ground_foot = np.max(np.array(inputs)[:, [31, 33]], axis=1)
-    rel_ground_foot = ground_foot - ground_foot[idx]
-    rel_ground_foot = rel_ground_foot.tolist()
-
-    # Order the people based on their distance
-    base = np.array([np.mean(np.array(keypoints[idx][0])), np.mean(np.array(keypoints[idx][1]))])
-    # delta_input = [abs((inp[31] + inp[33]) / 2 - base) for inp in inputs]
-    delta_input = [np.linalg.norm(base - np.array([np.mean(np.array(kp[0])), np.mean(np.array(kp[1]))]))
-                   for kp in keypoints]
-    sorted_indices = sorted(range(len(delta_input)), key=lambda k: delta_input[k])  # Return a list of sorted indices
-    all_inputs.extend(inputs[idx])
-
-    indices_idx = [idx for (idx, idx_gt) in matches]
-    if len(sorted_indices) > 2:
-        aa = 5
-    for ii in range(1, 3):
-        try:
-            index = sorted_indices[ii]
-
-            # Extract the idx_gt corresponding to the input we are attaching if it exists
-            try:
-                idx_idx_gt = indices_idx.index(index)
-                idx_gt = matches[idx_idx_gt][1]
-                all_inputs.append(rel_ground_foot[index])  # Relative lower ground foot
-                all_inputs.append(float(ys[idx_gt][3]))  # Output Z
-            except ValueError:
-                all_inputs.extend([0.] * 2)
-        except IndexError:
-            all_inputs.extend([0.] * 2)
-    assert len(all_inputs) == 34 + 2 * 2
-    return all_inputs
+def kps_apolloscape_to_monoloco(kps, kps_3d = False):
+    new_kps = []
+    for i in range(67):
+        if i not in kps[:,0]:
+            new_kps.append([i, -100, -100, -100])
+        else:
+            new_kps.append(kps[np.where(kps[:,0] == i)[0], :][0])
+    new_kps = np.array(new_kps)
+    
+    new_kps = new_kps[KPS_MAPPING] if kps_3d else new_kps[KPS_MAPPING][:,:3] #Remove or let the third dimension
+    
+    for i in range(len(new_kps)):
+        if new_kps[i,1] < 0:
+            new_kps[i, 0] = 0.0 # If the keypoint is not visible in the dataset, the confidence is of 0
+        else:
+            new_kps[i, 0] = 1.0 # If the keypoint is visible in the dataset, its confidence is of 1.0
+        
+    if kps_3d:
+        return new_kps[:,[1,2,3,0]]
+    else:
+        return new_kps[:,[1,2,0]]
 
 
-# def get_jean_yaw(box_obj):
-#     b_corners = box_obj.bottom_corners()
-#     center = box_obj.center
-#     back_point = [(b_corners[0, 2] + b_corners[0, 3]) / 2, (b_corners[2, 2] + b_corners[2, 3]) / 2]
-#
-#     x = b_corners[0, :] - back_point[0]
-#     y = b_corners[2, :] - back_point[1]
-#
-#     angle = math.atan2((x[0] + x[1]) / 2, (y[0] + y[1]) / 2) * 180 / 3.14
-#     angle = (angle + 360) % 360
-#     correction = math.atan2(center[0], center[2]) * 180 / 3.14
-#     return angle, correction
+
+def keypoints_to_cad_model(keypoints, vertices_cad_dic, radius = 160):
+    # Associate for each CAD model a set of keypoints
+    keypoints_to_cad = {}
+    num_keypoints = len(keypoints)
+    if len(keypoints.shape) == 1:
+        keypoints = [keypoints]
+    
+    
+    for i, vertices_2d in vertices_cad_dic.items():
+        
+        
+        for j, keypoint in enumerate(keypoints):
+            
+            res = np.ones([len(vertices_2d),2])*keypoint[1:] - np.transpose([vertices_2d[:,0]/vertices_2d[:,2], vertices_2d[:,1]/vertices_2d[:,2]])
+            dist = np.min( np.linalg.norm(res, axis = 1 ))
+            
+            if i == 0 and dist < radius:
+                keypoints_to_cad[j] = [i, dist]
+            elif i == 0 and dist >= radius:
+                keypoints_to_cad[j] = [-1, np.infty]
+                
+            elif dist < keypoints_to_cad[j][1] and dist <= radius:
+                keypoints_to_cad[j] = [i, dist]
+                
+    
+    count = 0
+    result = -1
+    
+    a = np.array( list( keypoints_to_cad.values()))[:,0]
+    index, counts = np.unique(a, return_counts=True)
+    
+    return keypoints_to_cad, index[np.argmax(counts)], np.max(counts)
+
+def keypoint_expander(vertices_2d, keypoints, buffer = 100) :
+    #process the existing keypoints and assign them a depth by minimizing the 
+    #distance with the vertices of the projected CAD models.
+    new_keypoints =[]
+    if len(keypoints.shape) == 1:
+        keypoints = [keypoints]
+        
+    for keypoint in keypoints:
+        if len(keypoint)<2:
+            print("ouch")
+        if len(vertices_2d[0])<2:
+            print("ouch 2")
+            
+            
+        res = np.ones([len(vertices_2d),2])*keypoint[1:] - np.transpose([vertices_2d[:,0]/vertices_2d[:,2], vertices_2d[:,1]/vertices_2d[:,2]])
+        
+        dist = np.linalg.norm(res, axis = 1)
+        
+        z = np.min(vertices_2d[np.argsort(dist)[0:buffer], 2])
+        
+        
+        new_keypoints.append([keypoint[0], keypoint[1], keypoint[2], z])
+        
+    return np.array(new_keypoints)
+
+def bbox_3d_extract(vertices):
+    
+    min_x = np.min(vertices[:,0])
+    max_x= np.max(vertices[:,0])
+    min_y = np.min(vertices[:,1])
+    max_y= np.max(vertices[:,1])
+    min_z = np.min(vertices[:,2])
+    max_z= np.max(vertices[:,2])
+    
+    h = max_y - min_y
+    w = max_x - min_x
+    l = max_z- min_z
+    
+    bbox_3d=[]
+    for x in [min_x, max_x]:
+        for y in [min_y, max_y]:
+            for z in [min_z, max_z]:
+                bbox_3d.append([x,y,z])
+                
+    
+    return(np.array(bbox_3d), w, l, h)
+
+def bbox_gt_extract(bbox_3d, kk):
+    zc = np.mean(bbox_3d[:,2])
+    
+    #take the top right corner and the bottom left corner of the bounding box in the 3D spac
+    corners_3d = np.array([[np.min(bbox_3d[:,0]), np.min(bbox_3d[:,1]), zc], [np.max(bbox_3d[:,0]), np.max(bbox_3d[:,1]), zc] ])
+    
+    box_2d = []
+    
+    for xyz in corners_3d:
+        xx, yy, zz = np.dot(kk, xyz)
+        uu = xx / zz
+        vv = yy / zz
+        box_2d.append(uu)
+        box_2d.append(vv)
+
+    return box_2d
+    
+def car_projection( car_model, scale , T, turn_over = False, bbox= False):
+    # perform the projection of the CAD models of the cars in the 3D space (use car_models_json)
+    with open(car_model) as json_file:
+        data = json.load(json_file) #Loading of the vehicle CAD model
+        vertices = np.array(data['vertices']) # extraction of the vertices
+
+        if (turn_over): # used to turn the vehicle over its z axis (180 degree turn) 
+                        # It was shown that the pose estimation have a precomputed rotation of the carts over their z axis 
+                        # Reminder: the z axis is the one of the depth
+            z_rot = np.array([0.,0., np.pi, 0.0,0.0,0.0])
+            vertices = project(z_rot, scale, vertices) #Pre-rotation of the vehicle over its z axis
+        
+        
+        vertices_r = project(T, scale, vertices) #projection in the 3D space
+        triangles = np.array(data['faces']) - 1      
+        results = vertices_r, triangles
+        
+        if bbox:
+            bbox_3d, w, l, h= bbox_3d_extract(vertices)
+            bbox_3d_r = project(T, scale, bbox_3d)      # Place the bounding box at its place in the 3D space
+            
+            results = vertices_r, triangles, bbox_3d_r, w, l, h
+        
+        return results
+
+
+
+def preprocess_monoloco_old(keypoints, kk, zero_center=False, kps_3d = False):
+
+    """ Preprocess batches of inputs
+    keypoints = torch tensors of (m, 3, 24)/(m,4,24)  or list [3,24]/[4,24]
+    Outputs =  torch tensors of (m, 48)/(m,72) in meters normalized (z=1) and zero-centered using the center of the box
+    
+    or, if we have the confidences:
+     Outputs =  torch tensors of (m, 72)/(m,96) in meters normalized (z=1) and zero-centered using the center of the box
+    """
+    if kps_3d:
+        nb_dim = 3
+    else: 
+        nb_dim = 2
+    
+    if isinstance(keypoints, list):
+        keypoints = torch.tensor(keypoints).double()
+    if isinstance(kk, list):
+        kk = torch.tensor(kk).double()
+    # Projection in normalized image coordinates and zero-center with the center of the bounding box
+    
+    xy1_all = pixel_to_camera(keypoints[:, 0:nb_dim, :], kk, 10)
+    
+    if zero_center:
+        uv_center = get_keypoints(keypoints, mode='center')
+        xy1_center = pixel_to_camera(uv_center, kk, 10)
+       
+    
+        kps_norm = xy1_all - xy1_center.unsqueeze(1)  # (m, 17, 3) - (m, 1, 3)
+    else:
+        kps_norm = xy1_all
+    kps_out = kps_norm[:, :, 0:nb_dim].reshape(kps_norm.size()[0], -1)  # no contiguous for view
+    #kps_out = torch.cat((kps_out, keypoints[:, nb_dim, :]), dim=1)
+    return kps_out
+
+
+# ## PIFPAF
+
+# In[10]:
+
+
+def pifpaf_info_extractor(json_pifpaf):
+    
+    #Extract the list of keypoints from the json files of openpifpaf
+    with open(json_pifpaf) as json_file:
+        data = json.load(json_file)
+        list_keypoints  =[]
+        for index, vehicle in enumerate(data):
+            #print("Vehicle n°", index)
+            keypoints = np.reshape(vehicle['keypoints'], [-1,3])[:,[2,0,1]] 
+            bbox = vehicle['bbox']
+            score = vehicle['score']
+            cat_id = vehicle['category_id']
+            
+            list_keypoints.append(keypoints)
+            
+        return list_keypoints
+
+
+class PreprocessApolloscape:
+
+
+    """Preprocess apolloscape dataset"""
+
+
+    dic_jo = {'train': dict(X=[], Y=[], names=[], kps=[], boxes_3d=[], K=[],
+                            clst=defaultdict(lambda: defaultdict(list))),
+              'val': dict(X=[], Y=[], names=[], kps=[], boxes_3d=[], K=[],
+                          clst=defaultdict(lambda: defaultdict(list))),
+                          }
+    dic_names = defaultdict(lambda: defaultdict(list))
+
+    def __init__(self, dir_ann, dataset, kps_3d = False, buffer=20, radius=200):
+
+        logging.basicConfig(level=logging.INFO)
+        self.logger = logging.getLogger(__name__)
+
+        self.buffer = buffer
+        self.radius = radius
+        
+        self.kps_3d = kps_3d
+        
+        self.dir_ann = dir_ann
+
+        dir_apollo = os.path.join('data', 'apolloscape')
+        dir_out = os.path.join('data', 'arrays')
+
+        assert os.path.exists(dir_apollo), "apollo directory does not exists"
+        assert os.path.exists(self.dir_ann), "The annotations directory does not exists"
+        assert os.path.exists(dir_out), "Joints directory does not exists"
+
+        now = datetime.datetime.now()
+        now_time = now.strftime("%Y%m%d-%H%M")[2:]
+        self.path_joints = os.path.join(dir_out, 'joints-apolloscape-' + dataset + '-' + now_time + '.json')
+        self.path_names = os.path.join(dir_out, 'names-apolloscape-' + dataset + '-' + now_time + '.json')
+
+        
+        self.path  = os.path.join(dir_apollo, dataset)
+        self.scenes, self.train_scenes, self.validation_scenes = factory(dataset, dir_apollo)
+    
+    
+    
+    def extract_ground_truth(self, car_poses,camera_id, scene_id):
+        
+        with open(car_poses) as json_file:
+            data = json.load(json_file) #open the pose of the cars
+
+        dic_vertices = {}
+        dic_boxes = {}
+        dic_poses = {}
+        vertices_to_keypoints = {}
+        dic_keypoints = {}
+
+        #Extract the boxes, vertices and the poses of each cars                   
+        for car_index, car in enumerate(data):
+            name_car = car_id2name[car['car_id']].name 
+            car_model = os.path.join(self.path, "car_models_json",name_car+".json")
+
+            intrinsic_matrix = intrinsic_vec_to_mat(K["Camera_"+camera_id])
+
+            vertices_r, triangles, bbox_3d , w, l, h = car_projection(car_model, np.array([1,1,1]), T = np.array(car['pose']),  turn_over = True, bbox = True
+                                                           )
+            vertices_2d = np.matmul(vertices_r,intrinsic_matrix.transpose()) # Projected vertices on the 2D plane
+            
+            box_gt = bbox_gt_extract(bbox_3d, intrinsic_matrix)  # Take the overal bounding box in the 2D space
+            
+            dic_vertices[car_index] = vertices_2d
+            dic_boxes[car_index] = [box_gt, w, l, h]
+            dic_poses[car_index] = np.array(car['pose'])
+
+        new_keypoints = None
+        
+        #  print("DIC_BOXES", dic_boxes)
+                           
+        if "sample" not in self.path:
+                           
+            car_keypoints = os.path.join(self.path, "keypoints", scene_id)
+            
+            keypoints_list = []
+            boxes_gt_list = []  # Countain the the 2D bounding box of the vehicles
+            boxes_3d_list = []
+            ys_list = []               
+            
+            #Compute the similarity between each set of car models in the 3D space and the set of keypoints in the 2D space
+            for index_keypoints, file in enumerate(os.listdir(car_keypoints)):
+                if file.endswith(".txt"):
+                    keypoints = np.loadtxt(os.path.join(car_keypoints, file))
+
+                    dic_keypoints[index_keypoints] = keypoints
+
+                    k_t_c, index_cad, count = keypoints_to_cad_model(keypoints, dic_vertices, radius = self.radius)
+
+                    if index_cad not in vertices_to_keypoints.keys():
+                        vertices_to_keypoints[index_cad] = [index_keypoints, count, k_t_c]
+                    elif vertices_to_keypoints[index_cad][1] < count:
+                        vertices_to_keypoints[index_cad] = [index_keypoints, count, k_t_c]
+            
+            for index_cad, (index_keypoints, count, k_t_c) in vertices_to_keypoints.items()   :
+
+                if (index_cad != -1 and count >=1):
+
+                    keypoints = dic_keypoints[index_keypoints]
+                    vertices_2d = dic_vertices[index_cad]
+                    new_keypoints = keypoint_expander(vertices_2d, keypoints, self.buffer)
+                    #Pixel keypoints with a Z component
+                    
+                    #Return an ordered list of the keypoints in the 3D space, their Bounding box in the 3D space and their pose in the 3D space
+                    keypoints_list.append( new_keypoints)
+                    
+                    
+                    boxes_gt_list.append(dic_boxes[index_cad][0]) #2D corners of the bounding box
+                    
+                    
+                    w, l, h = dic_boxes[index_cad][1:]
+                    pitch, yaw, roll, xc, yc, zc = dic_poses[index_cad] # Center position of the car and its orientation
+                    
+                    boxes_3d_list.append([xc, yc, zc, w, l, h])
+                    
+                    # CHECK IF IT IS PI or 2PI
+                    yaw = yaw%np.pi
+                    
+                    sin, cos, _ = correct_angle(yaw, [xc, yc, zc])
+                    
+                    if True :
+                        rtp = to_spherical([xc, yc, zc])
+                        r, theta, phi = rtp # With r =d = np.linalg.norm([xc,yc,zc]) -> conversion to spherical coordinates 
+                        ys_list.append([theta, phi, zc, r, h, w, l, sin, cos, yaw])
+                    else:
+                        ys_list.append([xc, yc, zc, np.linalg.norm([xc, yc, zc]), h, w, l, sin, cos, yaw])
+                    
+                    
+            
+            return boxes_gt_list, boxes_3d_list, keypoints_list, ys_list
+        
+        else:
+            
+            for index_cad, _ in dic_vertices.items() :
+                
+                boxes_gt_list.append(dic_boxes[index_cad][0]) #2D corners of the bounding box
+                    
+                w, l, h = dic_boxes[index_cad][1:]
+                pitch, yaw, roll, xc, yc, zc = dic_poses[index_cad] # Center position of the car and its orientation
+
+                boxes_3d_list.append([xc, yc, zc, w, l, h])
+                yaw = yaw%np.pi
+
+                ys_list.append([xc, yc, zc, np.linalg.norm([xc, yc, zc]), h, w, l, np.sin(yaw), np.cos(yaw), yaw])
+            
+            
+            return boxes_gt_list, boxes_3d_list, None, ys_lists
+    
+                    
+    def extract_ground_truth_pifpaf(self, car_poses,camera_id, scene_id, path_pif):
+        with open(car_poses) as json_file:
+            data = json.load(json_file) #open the pose of the cars
+
+        dic_vertices = {}
+        dic_boxes = {}
+        dic_poses = {}
+        vertices_to_keypoints = {}
+        dic_keypoints = {}
+
+        #Extract the boxes, vertices and the poses of each cars                   
+        for car_index, car in enumerate(data):
+            name_car = car_id2name[car['car_id']].name 
+            car_model = os.path.join(self.path, "car_models_json",name_car+".json")
+
+            intrinsic_matrix = intrinsic_vec_to_mat(K["Camera_"+camera_id])
+
+            vertices_r, triangles, bbox_3d , w, l, h = car_projection(car_model, np.array([1,1,1]), T = np.array(car['pose']),  turn_over = True, bbox = True
+                                                           )
+            vertices_2d = np.matmul(vertices_r,intrinsic_matrix.transpose()) # Projected vertices on the 2D plane
+            
+            box_gt = bbox_gt_extract(bbox_3d, intrinsic_matrix)  # Take the overal bounding box in the 2D space
+            
+            dic_vertices[car_index] = vertices_2d
+            dic_boxes[car_index] = [box_gt, w, l, h]
+            dic_poses[car_index] = np.array(car['pose'])
+
+        new_keypoints = None
+        
+        #  print("DIC_BOXES", dic_boxes)
+                           
+        if "sample" not in self.path:
+                           
+            car_keypoints = os.path.join(self.path, "keypoints", scene_id)
+            
+            keypoints_list = []
+            boxes_gt_list = []  # Countain the the 2D bounding box of the vehicles
+            boxes_3d_list = []
+            ys_list = []               
+            
+            keypoints_pifpaf = pifpaf_info_extractor(path_pif)
+            
+            #Compute the similarity between each set of car models in the 3D space and the set of keypoints from in the 2D space
+            for index_keypoints, keypoints in enumerate(keypoints_pifpaf):
+
+
+                dic_keypoints[index_keypoints] = keypoints
+
+                k_t_c, index_cad, count = keypoints_to_cad_model(keypoints, dic_vertices, radius = self.radius)
+
+                if index_cad not in vertices_to_keypoints.keys():
+                    vertices_to_keypoints[index_cad] = [index_keypoints, count, k_t_c]
+                elif vertices_to_keypoints[index_cad][1] < count:
+                    vertices_to_keypoints[index_cad] = [index_keypoints, count, k_t_c]
+            
+            for index_cad, (index_keypoints, count, k_t_c) in vertices_to_keypoints.items()   :
+
+                if (index_cad != -1 and count >=1):
+
+                    keypoints = dic_keypoints[index_keypoints]
+                    vertices_2d = dic_vertices[index_cad]
+                    new_keypoints = keypoint_expander(vertices_2d, keypoints, self.buffer)
+
+                    if self.kps_3d :
+                        keypoints_list.append( new_keypoints[:,[1,2,3,0]])
+                    else:
+                        keypoints_list.append( new_keypoints[:,[1,2,0]])
+                                        
+                    boxes_gt_list.append(dic_boxes[index_cad][0]) #2D corners of the bounding box
+                    
+                    
+                    w, l, h = dic_boxes[index_cad][1:]
+                    pitch, yaw, roll, xc, yc, zc = dic_poses[index_cad] # Center position of the car and its orientation
+                    
+                    boxes_3d_list.append([xc, yc, zc, w, l, h])
+                    
+                    # CHECK IF IT IS PI or 2PI
+                    yaw = yaw%np.pi
+                    
+                    sin, cos, _ = correct_angle(yaw, [xc, yc, zc])
+                    
+                    if True :
+                        rtp = to_spherical([xc, yc, zc])
+                        r, theta, psi = rtp # With r =d = np.linalg.norm([xc,yc,zc]) -> conversion to spherical coordinates 
+                        #print("THETA, PSI, R", theta, psi, r)
+                        ys_list.append([theta, psi, zc, r, h, w, l, sin, cos, yaw])
+                    else:
+                        ys_list.append([xc, yc, zc, np.linalg.norm([xc, yc, zc]), h, w, l, sin, cos, yaw])
+                    
+            return boxes_gt_list, boxes_3d_list, keypoints_list, ys_list
+        
+        
+        
+        else:
+            for index_cad, _ in dic_vertices.items() :
+                
+                boxes_gt_list.append(dic_boxes[index_cad][0]) #2D corners of the bounding box
+                    
+                w, l, h = dic_boxes[index_cad][1:]
+                pitch, yaw, roll, xc, yc, zc = dic_poses[index_cad] # Center position of the car and its orientation
+
+                boxes_3d_list.append([xc, yc, zc, w, l, h])
+                yaw = yaw%np.pi
+
+                ys_list.append([xc, yc, zc, np.linalg.norm([xc, yc, zc]), h, w, l, np.sin(yaw), np.cos(yaw), yaw])
+            
+            return boxes_gt_list, boxes_3d_list, None, ys_lists
+            
+    
+    
+    def run(self):
+        """
+        Prepare arrays for training
+        """
+        cnt_scenes  = cnt_ann = 0
+        start = time.time()
+        
+        for ii, scene in enumerate(self.scenes):
+            
+            if ii==-100:
+                print("BREAK")
+                break
+            
+            cnt_scenes +=1 
+            
+            scene_id = scene.split("/")[-1].split(".")[0]
+            camera_id = scene_id.split("_")[-1]
+            car_poses = os.path.join(self.path, "car_poses", scene_id + ".json")
+            
+            #print(scene_id, self.train_scenes[:5])
+            if scene_id+".jpg" in self.train_scenes:
+                phase = 'train'
+            elif scene_id+".jpg" in self.validation_scenes:
+                phase ='val'
+            else:    
+                print("phase name not in training or validation split")
+                continue
+                
+            kk = K["Camera_"+camera_id]#intrinsic_vec_to_mat( K["Camera_"+camera_id])
+            
+            path_im = scene
+            
+            # Run IoU with pifpaf detections and save
+            path_pif = os.path.join(self.dir_ann, scene_id+".jpg" + '.predictions.json')
+                        
+            if os.path.isfile(path_pif) and False:
+                boxes_gt_list, boxes_3d_list, kps_list, ys_list  = self.extract_ground_truth_pifpaf(car_poses,camera_id, scene_id, path_pif)
+            else:
+                boxes_gt_list, boxes_3d_list, kps_list, ys_list = self.extract_ground_truth(car_poses,camera_id, scene_id)
+                
+   
+            self.dic_names[scene_id+".jpg"]['boxes'] = copy.deepcopy(list(boxes_gt_list))
+            self.dic_names[scene_id+".jpg"]['ys'] = copy.deepcopy(ys_list)
+            self.dic_names[scene_id+".jpg"]['K'] = copy.deepcopy(intrinsic_vec_to_mat(kk).tolist())
+
+            
+            for kps, ys, boxes_gt, boxes_3d in zip(kps_list, ys_list, boxes_gt_list, boxes_3d_list):
+                
+                
+                kps = [kps.transpose().tolist()]
+                #kk = list(intrinsic_vec_to_mat(kk).transpose())
+                
+                self.dic_jo[phase]['kps'].append(kps)
+                
+                inp = preprocess_monoloco(kps,  intrinsic_vec_to_mat(kk).tolist(), kps_3d = self.kps_3d).view(-1).tolist()
+                #inp = preprocess_monoloco(kps,  intrinsic_mat_to_vec(intrinsic_vec_to_mat(kk).transpose()).tolist(), kps_3d = self.kps_3d).view(-1).tolist()
+                
+                #print("inp 2.0 \n", inp)
+                self.dic_jo[phase]['X'].append(inp)
+                self.dic_jo[phase]['Y'].append(list(ys))
+                self.dic_jo[phase]['names'].append(scene_id+".jpg")  # One image name for each annotation
+                self.dic_jo[phase]['boxes_3d'].append(list(boxes_3d))
+                self.dic_jo[phase]['K'].append(intrinsic_vec_to_mat(kk).tolist())
+                
+                append_cluster(self.dic_jo, phase, list(inp), list(ys), kps)
+                cnt_ann += 1
+                sys.stdout.write('\r' + 'Saved annotations {}'.format(cnt_ann) + '\t')
+
+        with open(os.path.join(self.path_joints), 'w') as f:
+            json.dump(self.dic_jo, f)
+        with open(os.path.join(self.path_names), 'w') as f:
+            json.dump(self.dic_names, f)
+        end = time.time()
+
+        extract_box_average(self.dic_jo['train']['boxes_3d'])
+        print("\nSaved {} annotations for {} scenes. Total time: {:.1f} minutes".format(cnt_ann, cnt_scenes, (end-start)/60))
+        print("\nOutput files:\n{}\n{}\n".format(self.path_names, self.path_joints))       
+                           
+
+            
+                    
+      
+
+
+# In[14]:
+
+
+def factory(dataset, dir_apollo):
+    """Define dataset type and split training and validation"""
+
+    assert dataset in ['train', '3d_car_instance_sample']
+
+    path = os.path.join(dir_apollo, dataset)
+    
+    if dataset == 'train':
+        
+        with open(os.path.join(path, "split", "train-list.txt"), "r") as file:
+            train_scenes = file.read().splitlines()
+        #scenes = [scene for scene in scenes if scene['token'] in train_scenes]
+        with open(os.path.join(path, "split", "validation-list.txt"), "r") as file:
+            validation_scenes = file.read().splitlines()
+            
+    elif dataset == '3d_car_instance_sample':
+        with open(os.path.join(path, "split", "train-list.txt"), "r") as file:
+            train_scenes = file.read().splitlines()
+        #scenes = [scene for scene in scenes if scene['token'] in train_scenes]
+        with open(os.path.join(path, "split", "validation-list.txt"), "r") as file:
+            validation_scenes = file.read().splitlines()
+    
+    path_img = os.path.join(path, "images")
+    scenes = [os.path.join(path_img, file) for file in os.listdir(path_img) if file.endswith(".jpg")]
+    
+    return scenes, train_scenes, validation_scenes
+
+
+
+#prep = PreprocessApolloscape(dir_ann, dir_apollo, dir_out, dataset, kps_3d = kps_3d, buffer = buffer, radius = radius)
+
+
+
+# 
+# Files with the first preprocess monoloco including the confidence for the keypoints:
+# 
+# inp = preprocess_monoloco(kps,  intrinsic_vec_to_mat(kk).transpose().tolist(), kps_3d = self.kps_3d).view(-1).tolist()
+#     
+# Output files:
+# /home/maximebonnesoeur/monstereo/data/arrays/names-train-201018-1910.json
+# /home/maximebonnesoeur/monstereo/data/arrays/joints-train-201018-1910.json
+
+# _____
+
+# In[ ]:
+
+
+
+
